refactor: clean up re-exports and import aliases
